requirements.txt
Pillow==9.5.0
numpy==1.24.3
opencv-python==4.7.0.72
matplotlib==3.7.1
scipy==1.10.1
scikit-image==0.20.0
